feat(day11): fast as I can make it
I considered a cache for count values but discounted it because I figured the chance
of cache hits would be small considering the key for the cache would be both the
stone and the iterations remaining.  In addition, the lookup times for the hash
map that I thought just the lookups would kill performance.

But looking online, it seems like this is the way people are solving it.